==> campus-leasing-company-reviews/01-convert-to-csv.ipynb <==
import pandas as pd
import numpy as np

df_byu_b = pd.read_excel('brigham-young-businesses.xlsx')
df_byu_r = pd.read_excel('brigham-young-reviews.xlsx')
df_psu_b = pd.read_excel('penn-state-businesses.xlsx')
df_psu_r = pd.read_excel('penn-state-reviews.xlsx')
df_uiuc_b = pd.read_excel('uiuc-businesses.xlsx')
df_uiuc_r = pd.read_excel('uiuc-reviews.xlsx')
df_iu_b = pd.read_excel('iu-businesses-sample.xlsx')
df_iu_r = pd.read_excel('iu-reviews-sample.xlsx')

display(df_byu_b.head(2))
display(df_byu_r.head(2))
display(df_psu_b.head(2))
display(df_psu_r.head(2))
display(df_uiuc_b.head(2))
display(df_uiuc_r.head(2))
display(df_iu_b.head(2))
display(df_iu_r.head(2))

df_byu_b.to_csv('byu-businesses.csv', index=None)
df_byu_r.to_csv('byu-reviews.csv', index=None)
df_psu_b.to_csv('psu-businesses.csv', index=None)
df_psu_r.to_csv('psu-reviews.csv', index=None)
df_uiuc_b.to_csv('uiuc-businesses.csv', index=None)
df_uiuc_r.to_csv('uiuc-reviews.csv', index=None)
df_iu_b.to_csv('iu-businesses-sample.csv', index=None)
df_iu_r.to_csv('iu-reviews-sample.csv', index=None)
