==> campus-leasing-company-reviews/02-clean-data.ipynb <==
# Clean Leasing Company Google Reviews Data

import pandas as pd
import numpy as np

pd.set_option('display.max_columns', 100)

## Load data

Read the Excel files. These are unmodified files from [Outscraper](https://outscraper.com/)'s scraped output. They contain Google reviews for leasing companies, apartment complexes, and student housing for three different campustowns.

- Brigham Young University (Provo, UT)
- Penn State University (University Park, PA)
- University of Illinois at Urbana Champaign (Champaign, IL)

df_byu_b = pd.read_csv('https://github.com/bdi475/datasets/raw/main/campustowns-leasing-company-reviews/byu-businesses.csv')
df_byu_r = pd.read_csv('https://github.com/bdi475/datasets/raw/main/campustowns-leasing-company-reviews/byu-reviews.csv')
df_psu_b = pd.read_csv('https://github.com/bdi475/datasets/raw/main/campustowns-leasing-company-reviews/psu-businesses.csv')
df_psu_r = pd.read_csv('https://github.com/bdi475/datasets/raw/main/campustowns-leasing-company-reviews/psu-reviews.csv')
df_uiuc_b = pd.read_csv('https://github.com/bdi475/datasets/raw/main/campustowns-leasing-company-reviews/uiuc-businesses.csv')
df_uiuc_r = pd.read_csv('https://github.com/bdi475/datasets/raw/main/campustowns-leasing-company-reviews/uiuc-reviews.csv')

display(df_byu_b.head(2))
display(df_byu_r.head(2))
display(df_psu_b.head(2))
display(df_psu_r.head(2))
display(df_uiuc_b.head(2))
display(df_uiuc_r.head(2))

## Preprocess data

### Add campus information column

Add campus information to each list of businesses dataframes. 

df_byu_b['campus'] = 'Brigham Young University'
df_psu_b['campus'] = 'Penn State University Park'
df_uiuc_b['campus'] = 'University of Illinois Urbana-Champaign'

### Concatenate data

# businesses
df_b = pd.concat([df_byu_b, df_psu_b, df_uiuc_b])

# reviews
df_r = pd.concat([df_byu_r, df_psu_r, df_uiuc_r])

print(f'df_b.shape={df_b.shape}')
print(f'df_r.shape={df_r.shape}')

### Select only useful columns

`df_b` (list of businesses) contains 61 columns and `df_r` contains 33 columns. Most of them are not useful for our analysis. Only select useful ones.

The columns you select can be different from the ones listed below.

df_b.head(3)

columns_to_select1 = [
    'campus',
    'place_id',
    'name',
    'site',
    'category',
    'borough',
    'street',
    'city',
    'postal_code',
    'state',
    'latitude',
    'longitude',
    'verified',
]

df_b = df_b[columns_to_select1].copy()

df_b.head()

df_r.head(3)

columns_to_select2 = [
    'place_id',
    'review_id',
    'author_id',
    'author_title',
    'review_text',
    'review_rating',
    'review_img_url',
    'review_datetime_utc',
    'owner_answer',
    'owner_answer_timestamp_datetime_utc',
    'review_likes'
]

df_r = df_r[columns_to_select2].copy()

df_r.head()

print(f'df_b.shape={df_b.shape}')
print(f'df_r.shape={df_r.shape}')

### Parse datetime columns

`review_datetime_utc` and `owner_answer_timestamp_datetime_utc` columns are currently `object`-typed (string types). Parse them as datetime using `pd.to_datetime()`.

df_r[['review_datetime_utc', 'owner_answer_timestamp_datetime_utc']].dtypes

df_r['review_datetime_utc'] = pd.to_datetime(df_r['review_datetime_utc'], utc=True)
df_r['owner_answer_timestamp_datetime_utc'] = pd.to_datetime(df_r['owner_answer_timestamp_datetime_utc'], utc=True)

Verify that the two datetime columns now display `datetime64` types.

df_r[['review_datetime_utc', 'owner_answer_timestamp_datetime_utc']].dtypes

df_r[['review_datetime_utc', 'owner_answer_timestamp_datetime_utc']].head()

## Save preprocessed data

Use the preprocessed files for Tableau visualizations.

df_b.to_csv('businesses.csv', index=None)
df_r.to_csv('reviews.csv', index=None)
