==> campus-leasing-company-reviews/03-EDA.ipynb <==
import pandas as pd
import numpy as np
import plotly.express as px

pd.set_option('display.max_columns', 100)

## Load data

Read the Excel files. These are unmodified files from Outscraper's scraped results.

The files contain Google reviews for leasing companies, apartment complexes, and student housing for three different campustowns.

- Brigham Young University (Provo, UT)
- Penn State University (University Park, PA)
- University of Illinois at Urbana Champaign (Champaign, IL)

df_b = pd.read_csv('businesses.csv')
df_r = pd.read_csv('reviews.csv', parse_dates = ['review_datetime_utc', 'owner_answer_timestamp_datetime_utc'])

display(df_b.head(2))
display(df_r.head(2))

df_r['author_id'].value_counts().head(10)

df_r[df_r['author_id'].isin(df_r['author_id'].value_counts().head(10).index)] \
    .drop(columns=['place_id', 'review_id', 'review_img_url']) \
    .sort_values('author_id')

df_r['year'] = df_r['review_datetime_utc'].dt.year
df_r['month'] = df_r['review_datetime_utc'].dt.month

df_r.head(2)
